work on some more redshift things
