+++ test/sql/redshift/aws_secret_redshift.test
