+++ test/sql/redshift/aws_secret_redshift_simple.test
@@ -1,0 +1,107 @@
+# name: test/sql/redshift/aws_secret_redshift_simple.test
+# description: test aws extension redshift secret (GetClusterCredentialsWithIAM) + postgres attach
+# group: [sql]
+
+require aws
+
+require parquet
+
+# postgres_scanner is built as a loadable (DONT_LINK) extension; `require` installs
+# it from the local build repository and loads it. `LOAD postgres` would not resolve
+# here because the bare name has no path when autoloading is disabled in the test build.
+require postgres_scanner
+
+# This test hits a real Redshift cluster + a real AWS identity, so it only runs
+# when the test cluster is available. The runner must already be authenticated to
+# AWS in a way that matches the CHAIN below (e.g. `aws sso login`).
+require-env REDSHIFT_TEST_SERVER_AVAILABLE 1
+
+statement ok
+SET allow_persistent_secrets=false
+
+# Creating this secret calls GetClusterCredentialsWithIAM and stores the minted
+# temporary db user/password alongside the connection details.
+# NOTE: adjust CHAIN/PROFILE to match how the runner authenticates to AWS.
+statement ok
+CREATE SECRET redshift_creds (
+    TYPE redshift,
+    PROVIDER credential_chain,
+    profile 'personal_not_root',
+    CHAIN 'config',
+    REGION 'eu-central-1',
+    REDSHIFT_CLUSTER_ID 'redshift-cluster-1',
+    REDSHIFT_DB_NAME 'dev',
+    REDSHIFT_HOST 'redshift-cluster-1.czzi0zkl6xg6.eu-central-1.redshift.amazonaws.com',
+    REDSHIFT_PORT '5439'
+);
+
+# The minted password must be redacted in the secret listing.
+query I
+SELECT secret_string LIKE '%redacted%' FROM (FROM duckdb_secrets() WHERE name = 'redshift_creds');
+----
+true
+
+
+# In this attach call, multiple API calls should be made to AWS redshift services.
+# 1 DescribeClusters with `aws redshift describe-clusters --cluster-identifier redshift-cluster-1`
+# See https://docs.aws.amazon.com/cli/latest/reference/redshift/describe-clusters.html
+# Returns the following (truncated for now)
+# {
+#     "Clusters": [
+#         {
+#             "ClusterIdentifier": "redshift-cluster-1",
+#             "NodeType": "rg.xlarge",
+#             "ClusterStatus": "paused",
+#             "ClusterAvailabilityStatus": "Paused",
+#             "MasterUsername": "duck",
+#             "DBName": "dev",
+#             "Endpoint": {
+#                 "Address": "redshift-cluster-1.czzi0zkl6xg6.eu-central-1.redshift.amazonaws.com",
+#                 "Port": 5439,
+#                 "VpcEndpoints": [
+#                     {
+#                         "VpcEndpointId": "vpce-0fbe3c2ae9f55dfab",
+#                         "VpcId": "vpc-31d6fd5a",
+#                         "NetworkInterfaces": [
+#                             {
+#                                 "NetworkInterfaceId": "eni-0a6385f01493b4b64",
+#                                 "SubnetId": "subnet-7361280e",
+#                                 "PrivateIpAddress": "172.31.39.157",
+#                                 "AvailabilityZone": "eu-central-1b"
+#                             }
+#                         ]
+#                     }
+#                 ]
+#             },
+#             "ClusterCreateTime": "2026-06-22T14:36:53.846000+00:00",
+#             "AutomatedSnapshotRetentionPeriod": 1,
+#             "ManualSnapshotRetentionPeriod": -1,
+#             "ClusterSecurityGroups": [],
+#             "VpcSecurityGroups": [
+#                 {
+#                     "VpcSecurityGroupId": "sg-9807d3f7",
+#                     "Status": "active"
+#                 }
+#             ],
+#             ...
+# }
+
+# Then it will call GetClusterCredentialsWithIAM to get a needed username and password
+statement ok
+ATTACH 'redshift-cluster-1' AS redshift_db (TYPE postgres, SECRET redshift_creds);
+
+statement ok
+select * from (show all tables) where database = 'redshift_db';
+
+
+# The dev database should expose exactly 7 tables.
+query I
+SELECT count(*) FROM (SHOW ALL TABLES) WHERE database = 'redshift_db';
+----
+7
+
+statement ok
+DETACH redshift_db
+
+statement ok
+DROP SECRET redshift_creds
